refactor: convert Builds to TS
