+++ src/index.js
@@ -7,8 +7,8 @@ const { Statistics } = require('./statistics');
 
 const { ActorClient } = require('./resource_clients/actor');
 const { ActorCollectionClient } = require('./resource_clients/actor_collection');
-const BuildClient = require('./resource_clients/build');
-// const BuildCollectionClient = require('./resource_clients/build_collection');
+const { BuildClient } = require('./resource_clients/build');
+// const { BuildCollectionClient } = require('./resource_clients/build_collection');
 const DatasetClient = require('./resource_clients/dataset');
 const DatasetCollectionClient = require('./resource_clients/dataset_collection');
 const KeyValueStoreClient = require('./resource_clients/key_value_store');
