--- src/resource_clients/build_collection.js
@@ -1,36 +1,0 @@
-const ow = require('ow').default;
-const { ResourceCollectionClient } = require('../base/resource_collection_client');
-
-/**
- * @hideconstructor
- */
-class BuildCollectionClient extends ResourceCollectionClient {
-    /**
-     * @param {ApiClientOptions} options
-     */
-    constructor(options) {
-        super({
-            resourcePath: options.resourcePath || 'actor-builds',
-            ...options,
-        });
-    }
-
-    /**
-     * https://docs.apify.com/api/v2#/reference/actors/build-collection/get-list-of-builds
-     * @param {object} [options]
-     * @param {number} [options.limit]
-     * @param {number} [options.offset]
-     * @param {boolean} [options.desc]
-     * @return {Promise<PaginationList>}
-     */
-    async list(options = {}) {
-        ow(options, ow.object.exactShape({
-            limit: ow.optional.number,
-            offset: ow.optional.number,
-            desc: ow.optional.boolean,
-        }));
-        return this._list(options);
-    }
-}
-
-module.exports = BuildCollectionClient;
